requirements.txt
# PyTorch (CPU)
torch==2.8.0+cpu
torchvision==0.23.0+cpu
torchaudio==2.8.0+cpu

# PyTorch Geometric dependencies (CPU)
torch-scatter==2.1.2
torch-sparse==0.7.4
torch-cluster==1.6.1
torch-spline-conv==1.3.1
torch-geometric==2.6.1

# Other utilities
numpy
pandas
Pillow
matplotlib
tqdm
requests
aiohttp
scikit-learn
streamlit
